refactor(workstation): vessel policy moves to a lifecycle-owning controller (#18460)
Workspace held the product-level vessel effects — open, close, park, reshow,
dispose — together with their receipts and the target-window state the park and
reshow bodies shared. That coupled a distinct concern's state, callers and
teardown to the composition root.

VesselController (Neo.controller.Component, beside the existing
WorkspaceController precedent) now owns the eight methods and seven state
fields. Workspace keeps four engine-required hooks, because the native
lifecycle binds keyFor/open/close off the workspace instance in registerSource
and resolves the tear-out override by prototype identity.

Six read-only getters stay on Workspace deliberately, and they are not
forwarding for moved members: DemoBWorkspace publishes the same receipt names
off its own workspace, so `getComponent(wsId, ['lastVesselParkReceipt'])` is a
cross-consumer observation convention. Relocating ownership must not move that
surface — which is why no E2E spec changed.

Two defects the move introduced and the tier caught:

  - `{windowId} = me` inside openTearOutVessel silently resolved to the
    CONTROLLER's windowId once `me` changed meaning. It did not throw; it sent
    null into the bootstrap read. Now `me.component`.
  - Diagnostic builders read receipts through `me.vesselController.X`, which
    throws under the prototype-call specs whose `this` is a stub. They read the
    guarded getter instead, so one path serves instance and stub.

Workspace 4,478 -> 3,899 lines; VesselController 763.
